+++ examples/oft-hyperliquid/deploy/MyHyperLiquidComposer_Recoverable.ts
@@ -1,10 +1,15 @@
 import assert from 'assert'
 
-import { Wallet } from 'ethers'
 import { type DeployFunction } from 'hardhat-deploy/types'
 import inquirer from 'inquirer'
 
-import { CHAIN_IDS, getCoreSpotDeployment, useBigBlock, useSmallBlock } from '@layerzerolabs/hyperliquid-composer'
+import {
+    CHAIN_IDS,
+    EthersSigner,
+    getCoreSpotDeployment,
+    useBigBlock,
+    useSmallBlock,
+} from '@layerzerolabs/hyperliquid-composer'
 
 const contractName_oft = 'MyOFT'
 const contractName_composer = 'HyperLiquidComposer_Recoverable'
@@ -35,7 +40,7 @@ const deploy: DeployFunction = async (hre) => {
     // Get logger from hardhat flag --log-level
     const loglevel = hre.hardhatArguments.verbose ? 'debug' : 'error'
 
-    const wallet = new Wallet(privateKey.toString())
+    const signer = new EthersSigner(privateKey.toString())
     const chainId = (await hre.ethers.provider.getNetwork()).chainId
     const isHyperliquid = chainId === CHAIN_IDS.MAINNET || chainId === CHAIN_IDS.TESTNET
     const isTestnet = chainId === CHAIN_IDS.TESTNET
@@ -81,7 +86,7 @@ const deploy: DeployFunction = async (hre) => {
 
     if (!isDeployed_composer) {
         console.log(`Switching to hyperliquid big block for the address ${deployer} to deploy ${contractName_composer}`)
-        const res = await useBigBlock(wallet, isTestnet, loglevel, true)
+        const res = await useBigBlock(signer, isTestnet, loglevel, true)
         console.log(res)
         console.log(`Deplying a contract uses big block which is mined at a transaction per minute.`)
     }
@@ -106,7 +111,7 @@ const deploy: DeployFunction = async (hre) => {
     // Set small block eitherway as we do not have a method to check which hyperliquidblock we are on
     {
         console.log(`Using small block with address ${deployer} for faster transactions`)
-        const res = await useSmallBlock(wallet, isTestnet, loglevel, true)
+        const res = await useSmallBlock(signer, isTestnet, loglevel, true)
         console.log(JSON.stringify(res, null, 2))
     }
 }