+++ examples/oapp-read/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @layerzerolabs/oapp-read-example
 
+## 0.3.3
+
+### Patch Changes
+
+- a2851bf: bump versions of monorepo packages
+
 ## 0.3.2
 
 ### Patch Changes