+++ packages/devtools/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @layerzerolabs/devtools
 
+## 2.0.4
+
+### Patch Changes
+
+- a2851bf: bump versions of monorepo packages
+- Updated dependencies [afb4dd3]
+  - @layerzerolabs/io-devtools@0.3.2
+
+## 2.0.3
+
+### Patch Changes
+
+- d720011: print max supply given solana local decimals
+
+## 2.0.2
+
+### Patch Changes
+
+- 06dc63b: quality of life improvements
+
 ## 2.0.1
 
 ### Patch Changes